Module fxbox_taxonomy::api

Public-facing API

The API for communicating with devices.

This API is provided as Traits to be implemented:

by the low-level layers of the FoxBox, including the adapters;

by test suites and tools that need to simulate connected devices.

In turn, this API is used to implement:

the public-facing REST and WebSocket API;

the rules API (ThinkerBell).

Structs

HubRequest
A request for one or more hubs.

InputRequest
A request for one or more input endpoints.

OutputRequest
A request for one or more output endpoints.

Period
An acceptable interval of time.

Value
A value that can be sent or received through this API. FIXME: Define.

WatchOptions
Options for watching an endpoint. FIXME: Define.

Enums

Error
An error produced by one of the APIs in this module.

Traits

API
The public API.

EndPointAPI
Endpoint-level API

HubAPI
Hub-level API