Enum fxbox_taxonomy::api::Error

pub enum Error {
    NoSuchNode(Id<NodeId>),
    NoSuchGet(Id<Get>),
    NoSuchSet(Id<Set>),
    TypeError,
}

An error produced by one of the APIs in this module.

Variants

NoSuchNode

There is no such node connected to the Foxbox, even indirectly.

NoSuchGet

There is no such input channel connected to the Foxbox, even indirectly.

NoSuchSet

There is no such output channel connected to the Foxbox, even indirectly.

TypeError

Attempting to set a value with the wrong type