Struct fxbox_taxonomy::api::HubRequest

pub struct HubRequest {
    pub id: Option<HubId>,
    pub tags: Vec<String>,
    pub inputs: Vec<InputRequest>,
    pub outputs: Vec<OutputRequest>,
}

A request for one or more hubs.

Fields

id

If Some(id), return only the hub with the corresponding id.

tags

Restrict results to hubs that have all the tags in tags.

inputs

Restrict results to hubs that have all the inputs in inputs.

outputs

Restrict results to hubs that have all the outputs in outputs.