Struct fxbox_taxonomy::api::InputRequest

pub struct InputRequest {
    pub id: Option<EndPointId>,
    pub parent: Option<HubId>,
    pub ancestor: Option<HubId>,
    pub tags: Vec<String>,
    pub kind: Option<ValueKind>,
    pub poll: Option<Period>,
    pub trigger: Option<Period>,
}

A request for one or more input endpoints.

Fields

id

If Some(id), return only the endpoint with the corresponding id.

parent

If Some(id), return only endpoints that are immediate children of hub id.

ancestor

If Some(id), return only endpoints that are descendants of hub id.

tags

Restrict results to endpoints that have all the tags in tags.

kind

If Some(k), restrict results to endpoints that produce values of kind k.

poll

If Some(r), restrict results to endpoints that support polling with the acceptable period.

trigger

If Some(r), restrict results to endpoints that support trigger with the acceptable period.