Module fxbox_taxonomy::devices

Metadata on devices

Structs

EndPoint
An endpoint represents a single place where data can enter or leave a device. Note that endpoints support either a single kind of input or a single kind of output. Devices that support both inputs or outputs, or several kinds of inputs, or several kinds of outputs, are represented as hubs containing several endpoints.

Hub
A hub represents a node to which several endpoints, as well as other hubs can be connected. The FoxBox is the root hub. Simple devices that can do a single thing (e.g. a button) are endpoints, while complex devices containing several sensors or effectors are also hubs, in which each sensor and each effector is an endpoint.

Input
An input operation available on an endpoint.

Output
An output operation available on an endpoint.

Enums

Type

ValueKind
The kind of value provided by an endpoint.

Type Definitions

EndPointId

HubId