Struct fxbox_taxonomy::devices::EndPoint

pub struct EndPoint<IO> {
    // some fields omitted
}

An endpoint represents a single place where data can enter or leave a device. Note that endpoints support either a single kind of input or a single kind of output. Devices that support both inputs or outputs, or several kinds of inputs, or several kinds of outputs, are represented as hubs containing several endpoints.

Methods

impl<IO> EndPoint<IO>

fn get_tags<'a>(&'a self) -> &'a Vec<String>

Tags describing the endpoint.

These tags can be set by the user, adapters or applications. They are used to regroup endpoints for rules.

For instance "entrance".

fn get_id<'a>(&'a self) -> &'a EndPointId

An id unique to this endpoint.

fn get_hub_id<'a>(&'a self) -> &'a HubId

The hub owning this endpoint.

fn get_mechanism<'a>(&'a self) -> &'a IO

The update mechanism for this endpoint.

fn get_last_seen(&self) -> DateTime<UTC>

The last time the device was seen.