Struct fxbox_taxonomy::devices::Hub

pub struct Hub {
    // some fields omitted
}

A hub represents a node to which several endpoints, as well as other hubs can be connected. The FoxBox is the root hub. Simple devices that can do a single thing (e.g. a button) are endpoints, while complex devices containing several sensors or effectors are also hubs, in which each sensor and each effector is an endpoint.

Methods

impl Hub

fn get_tags<'a>(&'a self) -> &'a Vec<String>

Tags describing the hub.

These tags can be set by the user, adapters or applications. They are used by applications.

For instance "entrance".

fn get_id<'a>(&'a self) -> &'a HubId

An id unique to this hub.

fn get_parent<'a>(&'a self) -> &'a Option<HubId>

If this hub has a parent, the id of the parent.

fn get_subhubs<'a>(&'a self) -> &'a Vec<Hub>

Hubs depending on this hub.

fn get_inputs<'a>(&'a self) -> &'a Vec<EndPoint<Input>>

Input endpoints connected directly to this hub.

fn get_outputs<'a>(&'a self) -> &'a Vec<EndPoint<Output>>

Output endpoints connected directly to this hub.