Struct fxbox_taxonomy::devices::Input

pub struct Input {
    // some fields omitted
}

An input operation available on an endpoint.

Methods

impl Input

fn get_kind(&self) -> ValueKind

The kind of value that can be obtained from this endpoint.

fn get_poll(&self) -> Option<Duration>

If Some(duration), this endpoint can be polled, i.e. it will respond when the FoxBox requests the latest value. Parameter duration indicates the smallest interval between two updates.

Otherwise, the endpoint cannot be polled and will push data to the FoxBox when it is available.

Examples

Long-running pollution or humidity sensors typically do not accept requests and rather send batches of data every 24h.

fn get_trigger(&self) -> Option<Duration>

If Some(duration), this endpoint can send the data to the FoxBox whenever it is updated. Parameter duration indicates the smallest interval between two updates.

Otherwise, the endpoint cannot send data to the FoxBox and needs to be polled.

fn get_updated(&self) -> DateTime<UTC>

Date at which the latest value was received, whether through polling or through a trigger.

Limitation

This is not a live view.