Struct fxbox_taxonomy::devices::Service

pub struct Service<IO> where IO: IOMechanism {
    pub tags: Vec<String>,
    pub id: ServiceId,
    pub node: NodeId,
    pub mechanism: IO,
    pub last_seen: Option<TimeStamp>,
    // some fields omitted
}

An service represents a single place where data can enter or leave a device. Note that services support either a single kind of input or a single kind of output. Devices that support both inputs or outputs, or several kinds of inputs, or several kinds of outputs, are represented as nodes containing several services.

Fields

tags

Tags describing the service.

These tags can be set by the user, adapters or applications. They are used to regroup services for rules.

For instance "entrance".

id

An id unique to this service.

node

The node owning this service.

mechanism

The update mechanism for this service.

last_seen

The last time the device was seen.