Module fxbox_taxonomy::values

Values that may be sent to/received from devices

Values manipulated by services

Structs

ExtNumeric
A data structure holding a numeric value of a type that has not been standardized yet.

Json
Representation of an object in JSON. It is often (albeit not always) possible to choose a more precise data structure for representing values send/accepted by a service. If possible, adapters should rather pick such more precise data structure.

TimeStamp

ValDuration

Enums

Color
A color. Internal representation may vary. The FoxBox adapters are expected to perform conversions to the format requested by their device.

Range
A comparison between two values.

Temperature
A temperature. Internal representation may be either Fahrenheit or Celcius. The FoxBox adapters are expected to perform conversions to the format requested by their devices.

Type
The type of values manipulated by endpoints.

Value
Representation of an actual value that can be sent to/received from a service.