Enum fxbox_taxonomy::values::Type

pub enum Type {
    Unit,
    Bool,
    Duration,
    TimeStamp,
    Temperature,
    String,
    Color,
    Json,
    Binary,
    ExtNumeric,
}

The type of values manipulated by endpoints.

Variants

Unit

Trivial values

An empty value. Used for instance to inform that a countdown has reached 0 or that a device is ready.

Bool

A boolean. Used for instance for on-off switches, presence detectors, etc.

Duration

Time

A duration. Used for instance in countdowns.

TimeStamp

A precise timestamp. Used for instance to determine when an event has taken place.

Temperature

String

Color

...

Json

Binary

ExtNumeric