Enum fxbox_taxonomy::values::Value

pub enum Value {
    Unit,
    Bool(bool),
    Duration(ValDuration),
    TimeStamp(TimeStamp),
    Temperature(Temperature),
    Color(Color),
    String(Arc<String>),
    ExtNumeric(ExtNumeric),
    Json(Arc<Json>),
    Binary {
        data: Arc<Vec<u8>>,
        mimetype: String,
    },
}

Representation of an actual value that can be sent to/received from a service.

Variants

Unit

Bool

Duration

TimeStamp

Temperature

Color

String

ExtNumeric

A numeric value representing a unit that has not been standardized yet into the API.

Json

A Json value. We put it behind an Arc to make sure that cloning remains unexpensive.

Binary

Binary data.

Fields

data

The actual data. We put it behind an Arc to make sure that cloning remains unexpensive.

mimetype

Methods

impl Value

fn get_type(&self) -> Type