Enum fxbox_taxonomy::api::Error

pub enum Error {
    NoSuchHub(HubId),
    NoSuchEndPoint(EndPointId),
    TypeError,
}

An error produced by one of the APIs in this module.

Variants

NoSuchHub

There is no such hub connected to the Foxbox, even indirectly.

NoSuchEndPoint

There is no such endpoint connected to the Foxbox, even indirectly.

TypeError

Attempting to set a value with the wrong type